+++ Window.Domain/ViewModels/Admin/ShopProduct/EditShopProductAdminSideDTO.cs
@@ -35,7 +35,7 @@ public record EditShopProductAdminSideDTO
 
     public bool IsInCustomersSuggestions { get; set; }
 
-    public int SaleRatio { get; set; }
+    public decimal SaleRatio { get; set; }
 
     #endregion
 }
