+++ Window.Domain/ViewModels/Seller/ShopProduct/CreateShopProductSellerSideDTO.cs
@@ -30,7 +30,7 @@ public class CreateShopProductSellerSideDTO
 
     public string Price { get; set; }
 
-    public int SaleRatio { get; set; }
+    public decimal SaleRatio { get; set; }
 
     #endregion
 }
